New Component: Data Iterator (DataTable abstraction) (#2297)
* Data Iterator behavior abstraction.  New VDataIterator component.  Refactor VDataTable.  Add new tests and fix old ones.

* Requested changes, moved expand functionality into mixin.  Change tag to content-tag, add content-props and content-class.  Kebab-case class name, move createProps into mixin, remove extraneous div on actions, add more tests and expectations.

* Add specific expectations to data table tests

* Addressed codebeat issue, simplified and optimized items rendering code

* Rebase: Move fixes into new locations

* Rebase: fix snapshots

* Fix snapshots, minor fix to list to correct empty class render

* Remove unnecessary spreads
