+++ src/components/index.js
@@ -12,6 +12,7 @@ export { default as VCard } from './VCard'
 export { default as VCarousel } from './VCarousel'
 export { default as VCheckbox } from './VCheckbox'
 export { default as VChip } from './VChip'
+export { default as VDataIterator } from './VDataIterator'
 export { default as VDataTable } from './VDataTable'
 export { default as VDatePicker } from './VDatePicker'
 export { default as VDialog } from './VDialog'
